requirements.txt
python-pptx==0.6.21
python-docx==0.8.11
markdown==3.4.4
google-generativeai==0.3.0
gradio==3.46.0
python-dotenv==1.0.0
